Cleanups, WIP on the plugin
