+++ gems/aws-sdk-cognitoidentity/lib/aws-sdk-cognitoidentity/client.rb
@@ -26,6 +26,7 @@ require 'aws-sdk-core/plugins/client_metrics_send_plugin.rb'
 require 'aws-sdk-core/plugins/transfer_encoding.rb'
 require 'aws-sdk-core/plugins/signature_v4.rb'
 require 'aws-sdk-core/plugins/protocols/json_rpc.rb'
+require 'aws-sdk-cognitoidentity/plugins/wrap_cognito_credentials.rb'
 
 Aws::Plugins::GlobalConfiguration.add_identifier(:cognitoidentity)
 
@@ -71,6 +72,7 @@ module Aws::CognitoIdentity
     add_plugin(Aws::Plugins::TransferEncoding)
     add_plugin(Aws::Plugins::SignatureV4)
     add_plugin(Aws::Plugins::Protocols::JsonRpc)
+    add_plugin(Aws::CognitoIdentity::Plugins::WrapCognitoCredentials)
 
     # @overload initialize(options)
     #   @param [Hash] options
