+++ gems/aws-sdk-core/lib/aws-sdk-core.rb
@@ -10,7 +10,6 @@ require_relative 'aws-sdk-core/credential_provider'
 require_relative 'aws-sdk-core/refreshing_credentials'
 require_relative 'aws-sdk-core/assume_role_credentials'
 require_relative 'aws-sdk-core/assume_role_web_identity_credentials'
-require_relative 'aws-sdk-core/cognito_identity_credentials'
 require_relative 'aws-sdk-core/credentials'
 require_relative 'aws-sdk-core/credential_provider_chain'
 require_relative 'aws-sdk-core/ecs_credentials'
