+++ services.json
@@ -146,7 +146,10 @@
     "models": "codestar-connections/2019-12-01"
   },
   "CognitoIdentity": {
-    "models": "cognito-identity/2014-06-30"
+    "models": "cognito-identity/2014-06-30",
+    "addPlugins": [
+      "Aws::CognitoIdentity::Plugins::WrapCognitoCredentials"
+    ]
   },
   "CognitoIdentityProvider": {
     "models": "cognito-idp/2016-04-18"
